增加User,Feed参数

+++ src/main/java/com/cuckoo/domain/UserMapper.java
@@ -11,13 +11,13 @@ public interface UserMapper {
     void addUser(User user);
 
     @Delete("DELETE FROM user_table WHERE id = #{id}")
-    void deleteUserById(Integer id);
+    void deleteUserById(Long id);
 
     @Select("SELECT * FROM user_table")
     List<User> getAllUsers();
 
     @Select("SELECT * FROM user_table WHERE id = #{id}")
-    User getUserById(Integer id);
+    User getUserById(Long id);
 
     @Select("SELECT * FROM user_table WHERE phone = #{phone}")
     User getUserByPhone(String phone);
@@ -26,6 +26,6 @@ public interface UserMapper {
     User getUserByUsername(String username);
 
     @Update("UPDATE user_table SET avatar=#{avatar}, updated = NOW() WHERE id=#{id}")
-    void updateAvatarById(@Param("id") Integer id,@Param("avatar") String avatar);
+    void updateAvatarById(@Param("id") Long id,@Param("avatar") String avatar);
 
 }
